Add fwd header for observable class
